requirements.txt
python-dotenv
requests
textblob
tweepy

